debug delete request9

+++ src/components/controlToken/checkWindow.js
@@ -6,12 +6,14 @@ const Load = window.addEventListener('load', ()=>{
 });
 
 const DeleteToken = window.addEventListener('beforeunload', async()=>{
-    const state = sessionStorage.getItem('userState');
-    const response = await fetch('https://toomuchstonestodo.onrender.com/userToken', {
-        method: 'DELETE',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({state})
-    });
+    if(sessionStorage.length){
+        const state = sessionStorage.getItem('userState');
+        const response = await fetch('https://toomuchstonestodo.onrender.com/userToken', {
+            method: 'DELETE',
+            headers: { 'Content-Type': 'application/json' },
+            body: JSON.stringify({state})
+        });
+    }
 });
 
 export default {Load, DeleteToken};
